Fix working version. TODO change clinkopts to relative path.
